21/1/24

+++ app/src/main/java/com/gathersg/user/CalendarAdapter.java
@@ -12,62 +12,47 @@ import androidx.recyclerview.widget.RecyclerView;
 import java.time.LocalDate;
 import java.util.ArrayList;
 
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-import java.time.LocalDate;
-import java.util.ArrayList;
-
-public class CalendarAdapter extends RecyclerView.Adapter<CalendarViewHolder>{
+public class CalendarAdapter extends RecyclerView.Adapter<CalendarViewHolder> {
     private final ArrayList<LocalDate> days;
     private final OnItemListener onItemListener;
 
-    public CalendarAdapter(ArrayList<LocalDate> days, OnItemListener onItemListener)
-    {
+    public CalendarAdapter(ArrayList<LocalDate> days, OnItemListener onItemListener) {
         this.days = days;
         this.onItemListener = onItemListener;
     }
 
     @NonNull
     @Override
-    public CalendarViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType)
-    {
+    public CalendarViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         LayoutInflater inflater = LayoutInflater.from(parent.getContext());
         View view = inflater.inflate(R.layout.fragment_calender_cell, parent, false);
         ViewGroup.LayoutParams layoutParams = view.getLayoutParams();
-        if(days.size()>15)
-            layoutParams.height =(int) (parent.getHeight()*0.166666666);
+        if (days.size() > 15)
+            layoutParams.height = (int) (parent.getHeight() * 0.166666666);
         else
-            layoutParams.height =(int) parent.getHeight();
+            layoutParams.height = parent.getHeight();
 
-        return new CalendarViewHolder(view, onItemListener,days);
+        return new CalendarViewHolder(view, onItemListener, days);
     }
 
     @Override
-    public void onBindViewHolder(@NonNull CalendarViewHolder holder, int position)
-    {
-       final LocalDate date = days.get(position);
-       if(date == null)
-           holder.dayOfMonth.setText("");
-       else {
-           holder.dayOfMonth.setText(String.valueOf(date.getDayOfMonth()));
-           if (date.equals(CalendarUtils.selectedDate))
-               holder.parentView.setBackgroundColor(Color.YELLOW);
-       }
+    public void onBindViewHolder(@NonNull CalendarViewHolder holder, int position) {
+        final LocalDate date = days.get(position);
+        if (date == null)
+            holder.dayOfMonth.setText("");
+        else {
+            holder.dayOfMonth.setText(String.valueOf(date.getDayOfMonth()));
+            if (date.equals(CalendarUtils.selectedDate))
+                holder.parentView.setBackgroundColor(Color.YELLOW);
+        }
     }
 
     @Override
-    public int getItemCount()
-    {
+    public int getItemCount() {
         return days.size();
     }
 
-    public interface  OnItemListener
-    {
+    public interface OnItemListener {
         void onItemClick(int position, LocalDate date);
 
     }
